# Conflicts:
#   Backend/Spring/src/main/java/coms309/controller/UserController.java
#   Backend/Spring/src/main/java/coms309/database/dataobjects/Assignment.java

+++ Backend/Spring/src/main/java/coms309/coderunner/JavaRunner.java
@@ -1,9 +1,15 @@
 package coms309.coderunner;
 
+import coms309.api.dataobjects.ApiAssignmentUnitTest;
 import coms309.api.dataobjects.ApiCodeSubmission;
 import coms309.database.dataobjects.AssignmentFile;
+import coms309.database.dataobjects.AssignmentUnitTest;
 
 import java.io.*;
+
+import java.util.LinkedList;
+import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.concurrent.TimeUnit;
 
 public class JavaRunner extends CodeRunner {
@@ -18,8 +24,8 @@ public class JavaRunner extends CodeRunner {
      * @param tfm TempFileManager object to manage temporary directory for executing file
      * @throws FileNotFoundException
      */
-    public JavaRunner(AssignmentFile af, ApiCodeSubmission acs, TempFileManager tfm) throws FileNotFoundException {
-        super(af.getCodeFolder(), tfm.getAssignmentFolderPath());
+    public JavaRunner(AssignmentFile af, ApiCodeSubmission acs, TempFileManager tfm, Iterable<AssignmentUnitTest> unitTests) throws FileNotFoundException {
+        super(af.getCodeFolder(), tfm.getAssignmentFolderPath(), unitTests);
         compiledRunner = true;
         this.fileName = acs.getName();
     }
@@ -33,29 +39,37 @@ public class JavaRunner extends CodeRunner {
      */
     @Override
     public boolean compile() throws IOException {
-        Process process = Runtime.getRuntime().exec("javac " + testFolder + "/" + fileName + " -d " + testFolder + "/out");
-        try {
-            if (process.waitFor(5, TimeUnit.SECONDS)) {
-                return process.exitValue() == 0;
-            }
-        } catch (InterruptedException e) {
-            return false;
+        ProcessBuilder processBuilder = new ProcessBuilder("javac", testFolder + "/" + fileName, "-d", testFolder + "/out");
+
+        ProcessManager processManager = new ProcessManager(processBuilder);
+        if(processManager.runForTime(5000)) {
+            stdOutData = processManager.getOutputData();
+            stdErrData = processManager.getErrorData();
+            return processManager.getExitValue() == 0;
         }
+        
+        processManager.terminateProcess();
+        
+        stdOutData = processManager.getOutputData();
+        stdErrData = processManager.getErrorData();
+
         return false;
     }
 
-    
-    /** 
-     * Method inherited from CodeRunner.  Runs the compiled Java program and saves the output/error streams so they can be checked later.
-     * 
-     * @return true if process runs successfully in allotted time with exit code 0, false otherwise
-     * @throws IOException
+    /**
+     * Method inherited from CodeRunner.  Runs the compiled Java program against the next unit test from the list, if available.
      */
-    @Override
-    public boolean run() throws IOException {
+    public AssignmentUnitTestResult runNext() throws IOException, NoSuchElementException {
+        if(!curTest.hasNext()) throw new NoSuchElementException("No more unit tests exist");
+
+        stdOutData = "";
+        stdErrData = "";
+
         String className = fileName.substring(0, fileName.indexOf('.'));
+
         Process process = Runtime.getRuntime().exec("java -cp " + testFolder + "/out " + className);
 
+        stdin = process.getOutputStream();
         stdout = process.getInputStream();
         stderr = process.getErrorStream();
 
@@ -63,6 +77,12 @@ public class JavaRunner extends CodeRunner {
 
         byte[] buff = new byte[1024];
 
+        AssignmentUnitTest aut = curTest.next();
+
+        stdin.write(aut.getInput().getBytes());
+        stdin.flush();
+
+        // run while process is alive and we have not hit a timeout
         while(process.isAlive() && System.currentTimeMillis() - startTime < maxRuntime) {
             while(stdout.available() > 0) {
                 int n = stdout.read(buff);
@@ -75,13 +95,66 @@ public class JavaRunner extends CodeRunner {
             }
         }
 
-
         if(process.isAlive()){
             process.destroyForcibly();
-            return false;
         }
 
-        return true;
+        return new AssignmentUnitTestResult(new ApiAssignmentUnitTest(aut), stdOutData, stdErrData, stdOutData.equals(aut.getExpectedOutput()));
+    }
+
+    
+    /** 
+     * Method inherited from CodeRunner.  Runs the compiled Java program and saves the output/error streams so they can be checked later.
+     * 
+     * @return true if process runs successfully in allotted time with exit code 0, false otherwise
+     * @throws IOException
+     */
+    @Override
+    public List<AssignmentUnitTestResult> run() throws IOException {
+        String className = fileName.substring(0, fileName.indexOf('.'));
+
+        List<AssignmentUnitTestResult> results = new LinkedList<>();
+
+        for (AssignmentUnitTest aut : unitTests) {
+
+            Process process = Runtime.getRuntime().exec("java -cp " + testFolder + "/out " + className);
+
+            stdin = process.getOutputStream();
+            stdout = process.getInputStream();
+            stderr = process.getErrorStream();
+
+            long startTime = System.currentTimeMillis();
+
+            byte[] buff = new byte[1024];
+
+            stdin.write(aut.getInput().getBytes());
+            stdin.flush();
+
+            // run while process is alive and we have not hit a timeout
+            while(process.isAlive() && System.currentTimeMillis() - startTime < maxRuntime) {
+                while(stdout.available() > 0) {
+                    int n = stdout.read(buff);
+                    stdOutData = stdOutData.concat(new String(buff, 0, n));
+                }
+
+                while(stderr.available() > 0) {
+                    int n = stderr.read(buff);
+                    stdErrData = stdErrData.concat(new String(buff, 0, n));
+                }
+            }
+
+            if(process.isAlive()){
+                process.destroyForcibly();
+            }
+
+            AssignmentUnitTestResult result = new AssignmentUnitTestResult(new ApiAssignmentUnitTest(aut), stdOutData, stdErrData, stdOutData.equals(aut.getExpectedOutput()));
+            results.add(result);
+
+            stdOutData = "";
+            stdErrData = "";
+        }
+
+        return results;
     }
 
 }
